remove middleware and rename routes
