+++ routes/thoughts.js
